third_party/WebKit/LayoutTests/storage/websql/change-version-expected.txt

CONSOLE ERROR: line 86: unable to open database (14 unable to open database file)
This test verifies that the JS database.changeVersion() function works as expected.
FAIL in openDatabase: InvalidStateError: Failed to execute 'openDatabase' on 'Window': unable to open database (14 unable to open database file)
